+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+### Version 2.0.0
+
+- Added ability to set attributes via method arguments during tags creation.
+- Added `addComment()` method for ability to add comment blocks.
+- Added methods `addText()` and `addHtml()` for adding plain text and html string into tags.
+- Method `append()`, `prepend()` were removed.
+
 ### Version 1.0.1
 
 Updated composer.json